Закончен переход на интервальные прогнозы
Полностью перешел на интервальные прогнозы, все тесты работают.
Следующая цель реализовать собиралку погоды для мейла.
И реализовать сравнивалку двух прогнозов по супер формуле.
И вообще продумать какие статистики мы будем считать.

+++ HistoryOfObservations/HistoryOfObservations/DayliObservation.cs
@@ -22,7 +22,7 @@ namespace HistoryOfObservations
             _tenDayTempIntervalPrediction = tenDayPredict;
         }
 
-        public int GetRealTempOfNoon()
+        public double GetRealTempOfNoon()
         {
             return _realTempOfADay.noonTemp;
         }
